Dogbone/prod_requirements.txt
amqp==5.0.6
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.86
botocore==1.20.86
cachetools==4.2.2
celery==5.0.5
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Django==2.2
django-bulk-update==2.2.0
django-constance==2.8.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-paypal==1.1.1
django-picklefield==3.0.1
django-queryinspect==1.1.0
django-redis-cache==3.0.0
django-storages==1.11.1
django-user-sessions==1.7.1
django-watchman==1.2.0
dropbox==11.10.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
fire==0.4.0
fonttools==4.24.4
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itsdangerous==2.0.1
jieba3k==0.35.1
jmespath==0.10.0
joblib==1.0.1
jsonfield==3.1.0
kombu==5.1.0
langdetect==1.0.9
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.19.0
oauth2client==4.1.3
opencv-python==4.5.2.52
packaging==20.9
pdf2docx==0.5.2
Pillow==8.2.0
ply==3.11
prompt-toolkit==3.0.18
protobuf==3.17.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyMuPDF==1.18.14
pyparsing==2.4.7
PyPDF2==1.26.0
PyReadableDiff==0.3.0
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.17.1
python-intercom==2.1.1
python-redis-lock==3.7.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.5.0
selenium==3.141.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stone==3.2.1
swapper==1.1.2.post1
termcolor==1.1.0
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
wikipedia==1.4.0